+++ apps/spotifun-rtk/src/common/components/Layout/Layout.tsx
@@ -1,7 +1,6 @@
 import type { ReactNode } from "react"
 import s from "./Layout.module.css"
-import { NavBar } from "@/common/components/NavBar/NavBar.tsx"
-
+import { NavBar } from "@/common/components"
 
 type Props = {
   children: ReactNode
@@ -10,7 +9,7 @@ type Props = {
 export const Layout = ({ children }: Props) => {
   return (
     <div className={s.layout}>
-      <NavBar/>
+      <NavBar />
       <div className={s.layoutContent}>{children}</div>
     </div>
   )
