+++ apps/spotifun-rtk/src/common/components/index.ts
@@ -6,5 +6,6 @@ export { PageNotFound } from "./PageNotFound/PageNotFound.tsx"
 export { PageTitle } from "./PageTitle/PageTitle.tsx"
 export { Pagination } from "./Pagination/Pagination.tsx"
 export { SearchInput } from "./SearchInput/SearchInput.tsx"
-export { Modal } from './Modal/Modal.tsx'
-export { IconButton } from './IconButton/IconButton.tsx'
+export { Modal } from "./Modal/Modal.tsx"
+export { IconButton } from "./IconButton/IconButton.tsx"
+export { NavBar } from "./NavBar/NavBar.tsx"
